feat: Add i18n support for Laser Tag page
