+++ i18n/messages/fr.json
@@ -36,6 +36,30 @@
       }
     }
   },
+  "laserTag": {
+    "tagline": "Le laser tag portable et immersif pour les événements japonais.",
+    "imageAlt": "Delta Quest VR Laser Tag",
+    "concept": {
+      "title": "Concept",
+      "description": "Delta Quest est une expérience de laser tag portable et immersive qui peut être intégrée dans tous les types de grands événements au Japon, tels que les tournois d'e-sport, les événements en stade ou les conventions d'anime et de jeux. Nous offrons une expérience de jeu « réelle » sans précédent grâce à la réalité virtuelle et aux collaborations avec des IPs."
+    },
+    "features": {
+      "title": "Caractéristiques",
+      "list": {
+        "portable": "Installation portable : S'adapte de manière flexible à tous les lieux d'événements.",
+        "ipCollaboration": "Expérience de marque par collaboration IP : Interagit avec les fans en s'associant à des animes et des jeux.",
+        "competitive": "Compétitif et familial : Le jeu propose un système de score, des matchs en équipe, et peut être apprécié par les parents et leurs enfants."
+      }
+    },
+    "targetEvents": {
+      "title": "Événements cibles",
+      "description": "Parfait pour les tournois d'e-sport, les événements sportifs, les conventions d'anime et de jeux, ainsi que de nombreux autres événements à grande échelle. C'est l'attraction idéale pour les organisateurs qui souhaitent augmenter l'engagement des visiteurs."
+    },
+    "contact": {
+      "title": "Contactez-nous",
+      "description": "Pour toute question ou demande de collaboration concernant ce service, n'hésitez pas à nous contacter."
+    }
+  },
   "home": {
     "welcome": "Bienvenue chez Delta Quest",
     "description": "Delta Quest fournit des services IT de pointe, du développement frontend à l'optimisation SEO et à l'infrastructure cloud.",
